Refactor velox/common/tests
