+++ velox/common/CMakeLists.txt
@@ -18,7 +18,4 @@ add_subdirectory(file)
 add_subdirectory(memory)
 add_subdirectory(process)
 add_subdirectory(serialization)
-if(${VELOX_BUILD_TESTING})
-  add_subdirectory(tests)
-endif()
 add_subdirectory(time)
