+++ velox/common/memory/tests/MappedMemoryTest.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 #include "velox/common/memory/MappedMemory.h"
-#include "velox/common/base/test_utils/GTestUtils.h"
+#include "velox/common/base/tests/GTestUtils.h"
 #include "velox/common/memory/MmapAllocator.h"
 
 #include <thread>
