+++ velox/common/process/CMakeLists.txt
@@ -15,3 +15,7 @@
 add_library(velox_process ProcessBase.cpp StackTrace.cpp TraceContext.cpp)
 
 target_link_libraries(velox_process ${FOLLY_WITH_DEPENDENCIES} glog::glog)
+
+if(${VELOX_BUILD_TESTING})
+  add_subdirectory(tests)
+endif()
