+++ velox/dwio/dwrf/test/ColumnWriterIndexTests.cpp
@@ -16,7 +16,7 @@
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include "velox/common/base/test_utils/GTestUtils.h"
+#include "velox/common/base/tests/GTestUtils.h"
 #include "velox/dwio/dwrf/test/OrcTest.h"
 #include "velox/dwio/dwrf/writer/ColumnWriter.h"
 #include "velox/vector/ComplexVector.h"
