+++ velox/dwio/dwrf/test/TestReader.cpp
@@ -18,7 +18,7 @@
 #include <gtest/gtest.h>
 #include "folly/Random.h"
 #include "folly/lang/Assume.h"
-#include "velox/common/base/test_utils/GTestUtils.h"
+#include "velox/common/base/tests/GTestUtils.h"
 #include "velox/common/caching/DataCache.h"
 #include "velox/dwio/common/DataSink.h"
 #include "velox/dwio/common/MemoryInputStream.h"
