+++ velox/vector/tests/BiasVectorTest.cpp
@@ -16,7 +16,7 @@
 
 #include <gtest/gtest.h>
 
-#include "velox/common/base/test_utils/GTestUtils.h"
+#include "velox/common/base/tests/GTestUtils.h"
 #include "velox/vector/SimpleVector.h"
 #include "velox/vector/tests/VectorMaker.h"
